Delete states updated again

+++ 0x0F-python-object_relational_mapping/13-model_state_delete_a.py
@@ -18,8 +18,7 @@ if __name__ == "__main__":
     Session = sessionmaker(bind=engine)
     session = Session()
 
-    states = session.query(State).filter(State.name.contains('a'))
-
-    for state in states:
-        session.delete(state)
-        session.commit()
+    for state in session.query(State):
+        if "a" in state.name:
+            session.delete(state)
+    session.commit()
